# BookingMovieTickets/Frontend/public/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "BookingMovieTickets",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Movie/{id}": {
      "get": {
        "tags": [
          "Movie"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json;odata.metadata=minimal;odata.streaming=true": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json;odata.metadata=minimal;odata.streaming=false": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json;odata.metadata=minimal": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json;odata.metadata=full;odata.streaming=true": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json;odata.metadata=full;odata.streaming=false": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json;odata.metadata=full": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json;odata.metadata=none;odata.streaming=true": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json;odata.metadata=none;odata.streaming=false": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json;odata.metadata=none": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json;odata.streaming=true": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json;odata.streaming=false": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json;odata.metadata=minimal;odata.streaming=true;IEEE754Compatible=false": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json;odata.metadata=minimal;odata.streaming=true;IEEE754Compatible=true": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json;odata.metadata=minimal;odata.streaming=false;IEEE754Compatible=false": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json;odata.metadata=minimal;odata.streaming=false;IEEE754Compatible=true": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json;odata.metadata=minimal;IEEE754Compatible=false": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json;odata.metadata=minimal;IEEE754Compatible=true": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json;odata.metadata=full;odata.streaming=true;IEEE754Compatible=false": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json;odata.metadata=full;odata.streaming=true;IEEE754Compatible=true": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json;odata.metadata=full;odata.streaming=false;IEEE754Compatible=false": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json;odata.metadata=full;odata.streaming=false;IEEE754Compatible=true": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json;odata.metadata=full;IEEE754Compatible=false": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json;odata.metadata=full;IEEE754Compatible=true": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json;odata.metadata=none;odata.streaming=true;IEEE754Compatible=false": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json;odata.metadata=none;odata.streaming=true;IEEE754Compatible=true": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json;odata.metadata=none;odata.streaming=false;IEEE754Compatible=true": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json;odata.metadata=none;odata.streaming=false;IEEE754Compatible=false": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json;odata.metadata=none;IEEE754Compatible=false": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json;odata.metadata=none;IEEE754Compatible=true": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json;odata.streaming=true;IEEE754Compatible=false": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json;odata.streaming=true;IEEE754Compatible=true": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json;odata.streaming=false;IEEE754Compatible=false": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json;odata.streaming=false;IEEE754Compatible=true": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json;IEEE754Compatible=false": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json;IEEE754Compatible=true": {
                "schema": {
                  "type": "string"
                }
              },
              "application/xml": {
                "schema": {
                  "type": "string"
                }
              },
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/octet-stream": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Movie"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json;odata.metadata=minimal;odata.streaming=true": {
              "schema": {
                "type": "string"
              }
            },
            "application/json;odata.metadata=minimal;odata.streaming=false": {
              "schema": {
                "type": "string"
              }
            },
            "application/json;odata.metadata=minimal": {
              "schema": {
                "type": "string"
              }
            },
            "application/json;odata.metadata=full;odata.streaming=true": {
              "schema": {
                "type": "string"
              }
            },
            "application/json;odata.metadata=full;odata.streaming=false": {
              "schema": {
                "type": "string"
              }
            },
            "application/json;odata.metadata=full": {
              "schema": {
                "type": "string"
              }
            },
            "application/json;odata.metadata=none;odata.streaming=true": {
              "schema": {
                "type": "string"
              }
            },
            "application/json;odata.metadata=none;odata.streaming=false": {
              "schema": {
                "type": "string"
              }
            },
            "application/json;odata.metadata=none": {
              "schema": {
                "type": "string"
              }
            },
            "application/json;odata.streaming=true": {
              "schema": {
                "type": "string"
              }
            },
            "application/json;odata.streaming=false": {
              "schema": {
                "type": "string"
              }
            },
            "application/json": {
              "schema": {
                "type": "string"
              }
            },
            "application/json;odata.metadata=minimal;odata.streaming=true;IEEE754Compatible=false": {
              "schema": {
                "type": "string"
              }
            },
            "application/json;odata.metadata=minimal;odata.streaming=true;IEEE754Compatible=true": {
              "schema": {
                "type": "string"
              }
            },
            "application/json;odata.metadata=minimal;odata.streaming=false;IEEE754Compatible=false": {
              "schema": {
                "type": "string"
              }
            },
            "application/json;odata.metadata=minimal;odata.streaming=false;IEEE754Compatible=true": {
              "schema": {
                "type": "string"
              }
            },
            "application/json;odata.metadata=minimal;IEEE754Compatible=false": {
              "schema": {
                "type": "string"
              }
            },
            "application/json;odata.metadata=minimal;IEEE754Compatible=true": {
              "schema": {
                "type": "string"
              }
            },
            "application/json;odata.metadata=full;odata.streaming=true;IEEE754Compatible=false": {
              "schema": {
                "type": "string"
              }
            },
            "application/json;odata.metadata=full;odata.streaming=true;IEEE754Compatible=true": {
              "schema": {
                "type": "string"
              }
            },
            "application/json;odata.metadata=full;odata.streaming=false;IEEE754Compatible=false": {
              "schema": {
                "type": "string"
              }
            },
            "application/json;odata.metadata=full;odata.streaming=false;IEEE754Compatible=true": {
              "schema": {
                "type": "string"
              }
            },
            "application/json;odata.metadata=full;IEEE754Compatible=false": {
              "schema": {
                "type": "string"
              }
            },
            "application/json;odata.metadata=full;IEEE754Compatible=true": {
              "schema": {
                "type": "string"
              }
            },
            "application/json;odata.metadata=none;odata.streaming=true;IEEE754Compatible=false": {
              "schema": {
                "type": "string"
              }
            },
            "application/json;odata.metadata=none;odata.streaming=true;IEEE754Compatible=true": {
              "schema": {
                "type": "string"
              }
            },
            "application/json;odata.metadata=none;odata.streaming=false;IEEE754Compatible=true": {
              "schema": {
                "type": "string"
              }
            },
            "application/json;odata.metadata=none;odata.streaming=false;IEEE754Compatible=false": {
              "schema": {
                "type": "string"
              }
            },
            "application/json;odata.metadata=none;IEEE754Compatible=false": {
              "schema": {
                "type": "string"
              }
            },
            "application/json;odata.metadata=none;IEEE754Compatible=true": {
              "schema": {
                "type": "string"
              }
            },
            "application/json;odata.streaming=true;IEEE754Compatible=false": {
              "schema": {
                "type": "string"
              }
            },
            "application/json;odata.streaming=true;IEEE754Compatible=true": {
              "schema": {
                "type": "string"
              }
            },
            "application/json;odata.streaming=false;IEEE754Compatible=false": {
              "schema": {
                "type": "string"
              }
            },
            "application/json;odata.streaming=false;IEEE754Compatible=true": {
              "schema": {
                "type": "string"
              }
            },
            "application/json;IEEE754Compatible=false": {
              "schema": {
                "type": "string"
              }
            },
            "application/json;IEEE754Compatible=true": {
              "schema": {
                "type": "string"
              }
            },
            "application/xml": {
              "schema": {
                "type": "string"
              }
            },
            "text/plain": {
              "schema": {
                "type": "string"
              }
            },
            "text/json": {
              "schema": {
                "type": "string"
              }
            },
            "application/*+json": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Movie"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Movie": {
      "post": {
        "tags": [
          "Movie"
        ],
        "requestBody": {
          "content": {
            "application/json;odata.metadata=minimal;odata.streaming=true": {
              "schema": {
                "type": "string"
              }
            },
            "application/json;odata.metadata=minimal;odata.streaming=false": {
              "schema": {
                "type": "string"
              }
            },
            "application/json;odata.metadata=minimal": {
              "schema": {
                "type": "string"
              }
            },
            "application/json;odata.metadata=full;odata.streaming=true": {
              "schema": {
                "type": "string"
              }
            },
            "application/json;odata.metadata=full;odata.streaming=false": {
              "schema": {
                "type": "string"
              }
            },
            "application/json;odata.metadata=full": {
              "schema": {
                "type": "string"
              }
            },
            "application/json;odata.metadata=none;odata.streaming=true": {
              "schema": {
                "type": "string"
              }
            },
            "application/json;odata.metadata=none;odata.streaming=false": {
              "schema": {
                "type": "string"
              }
            },
            "application/json;odata.metadata=none": {
              "schema": {
                "type": "string"
              }
            },
            "application/json;odata.streaming=true": {
              "schema": {
                "type": "string"
              }
            },
            "application/json;odata.streaming=false": {
              "schema": {
                "type": "string"
              }
            },
            "application/json": {
              "schema": {
                "type": "string"
              }
            },
            "application/json;odata.metadata=minimal;odata.streaming=true;IEEE754Compatible=false": {
              "schema": {
                "type": "string"
              }
            },
            "application/json;odata.metadata=minimal;odata.streaming=true;IEEE754Compatible=true": {
              "schema": {
                "type": "string"
              }
            },
            "application/json;odata.metadata=minimal;odata.streaming=false;IEEE754Compatible=false": {
              "schema": {
                "type": "string"
              }
            },
            "application/json;odata.metadata=minimal;odata.streaming=false;IEEE754Compatible=true": {
              "schema": {
                "type": "string"
              }
            },
            "application/json;odata.metadata=minimal;IEEE754Compatible=false": {
              "schema": {
                "type": "string"
              }
            },
            "application/json;odata.metadata=minimal;IEEE754Compatible=true": {
              "schema": {
                "type": "string"
              }
            },
            "application/json;odata.metadata=full;odata.streaming=true;IEEE754Compatible=false": {
              "schema": {
                "type": "string"
              }
            },
            "application/json;odata.metadata=full;odata.streaming=true;IEEE754Compatible=true": {
              "schema": {
                "type": "string"
              }
            },
            "application/json;odata.metadata=full;odata.streaming=false;IEEE754Compatible=false": {
              "schema": {
                "type": "string"
              }
            },
            "application/json;odata.metadata=full;odata.streaming=false;IEEE754Compatible=true": {
              "schema": {
                "type": "string"
              }
            },
            "application/json;odata.metadata=full;IEEE754Compatible=false": {
              "schema": {
                "type": "string"
              }
            },
            "application/json;odata.metadata=full;IEEE754Compatible=true": {
              "schema": {
                "type": "string"
              }
            },
            "application/json;odata.metadata=none;odata.streaming=true;IEEE754Compatible=false": {
              "schema": {
                "type": "string"
              }
            },
            "application/json;odata.metadata=none;odata.streaming=true;IEEE754Compatible=true": {
              "schema": {
                "type": "string"
              }
            },
            "application/json;odata.metadata=none;odata.streaming=false;IEEE754Compatible=true": {
              "schema": {
                "type": "string"
              }
            },
            "application/json;odata.metadata=none;odata.streaming=false;IEEE754Compatible=false": {
              "schema": {
                "type": "string"
              }
            },
            "application/json;odata.metadata=none;IEEE754Compatible=false": {
              "schema": {
                "type": "string"
              }
            },
            "application/json;odata.metadata=none;IEEE754Compatible=true": {
              "schema": {
                "type": "string"
              }
            },
            "application/json;odata.streaming=true;IEEE754Compatible=false": {
              "schema": {
                "type": "string"
              }
            },
            "application/json;odata.streaming=true;IEEE754Compatible=true": {
              "schema": {
                "type": "string"
              }
            },
            "application/json;odata.streaming=false;IEEE754Compatible=false": {
              "schema": {
                "type": "string"
              }
            },
            "application/json;odata.streaming=false;IEEE754Compatible=true": {
              "schema": {
                "type": "string"
              }
            },
            "application/json;IEEE754Compatible=false": {
              "schema": {
                "type": "string"
              }
            },
            "application/json;IEEE754Compatible=true": {
              "schema": {
                "type": "string"
              }
            },
            "application/xml": {
              "schema": {
                "type": "string"
              }
            },
            "text/plain": {
              "schema": {
                "type": "string"
              }
            },
            "text/json": {
              "schema": {
                "type": "string"
              }
            },
            "application/*+json": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "JWT Authentication for Booking Movie Tickets",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}